Move per-wizard serializers into their components
